Release APK for Windows
I faced the problem where I was trying to run './gradlew assembleRelease' and finally I found a solution that I must use ''gradlew assembleRelease' on windows. It would be better to mention that here in the documentation. So that developers don't have to go and search for this in other places. Thank you!